- реализован мэпинг на /flux/authors с помощью webflux

+++ src/main/java/otus/library/rest/AuthorController.java
@@ -1,64 +1,38 @@
 package otus.library.rest;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.server.RouterFunction;
+import org.springframework.web.reactive.function.server.ServerResponse;
 import otus.library.domain.Author;
 import otus.library.repository.AuthorRepository;
 
-import java.util.List;
-import java.util.Optional;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
+import static org.springframework.web.reactive.function.server.RouterFunctions.route;
+import static org.springframework.web.reactive.function.server.ServerResponse.ok;
 
-@Controller
+@Configuration
 public class AuthorController {
-    private final AuthorRepository authorRepository;
-
-    @Autowired
-    public AuthorController(AuthorRepository authorRepository) {
-        this.authorRepository = authorRepository;
-    }
-
-    @GetMapping("/authors")
-    public String listPage(Model model) {
-        List<Author> authors = authorRepository.findAll();
-        model.addAttribute("authors", authors);
-        return "authors/list";
-    }
-
-    @GetMapping("/authors/edit")
-    public String editPage(@RequestParam("id") String id, Model model) {
-        Author authors = authorRepository.findById(id).orElseThrow(NotFoundException::new);
-        model.addAttribute("authors", authors);
-        return "authors/edit";
-    }
-
-    @PostMapping("/authors/create")
-    public String createAuthor(@RequestParam("lname") String lname, @RequestParam("fname") String fname, Model model) {
-        authorRepository.save(new Author(fname, lname));
-        model.addAttribute("backref", "/authors");
-        return "save";
-    }
-
-    @PostMapping("/authors/delete")
-    public String deleteAuthor(@RequestParam("id") String id, Model model) {
-        authorRepository.deleteById(id);
-        model.addAttribute("backref", "/authors");
-        return "save";
-    }
-
-    @GetMapping("/authors/new")
-    public String newPage() {
-        return "authors/new";
-    }
-
-    @PostMapping("/authors/save")
-    public String saveAuthor(@RequestParam("id") String id, @RequestParam("lname") String lname, @RequestParam("fname") String fname, Model model) {
-        Optional<Author> author = authorRepository.findById(id);
-        author.get().setFname(fname);
-        author.get().setLname(lname);
-        authorRepository.save(author.get());
-        model.addAttribute("backref", "/authors");
-        return "save";
+    @Bean
+    public RouterFunction<ServerResponse> authorsRoutes(AuthorRepository authorRepository) {
+        return route()
+                .GET("/flux/authors", request -> ok().contentType(APPLICATION_JSON).body(authorRepository.findAll(), Author.class))
+                .DELETE("/flux/authors/{id}", request -> authorRepository.deleteById(request.pathVariable("id"))
+                        .then(ok().build()))
+                .PUT("/flux/authors/{id}", request -> {
+                            Author author = new Author();
+                            author.setId(request.pathVariable("id"));
+                            author.setFname(request.queryParam("fname").get());
+                            author.setLname(request.queryParam("lname").get());
+                            return ok().contentType(APPLICATION_JSON).body(authorRepository.save(author), Author.class);
+                        }
+                )
+                .POST("/flux/authors",
+                        request -> ok().contentType(APPLICATION_JSON).body(authorRepository.save(
+                                new Author(request.queryParam("fname").get(), request.queryParam("lname").get())),
+                                Author.class
+                            )
+                )
+                .build();
     }
 }
